+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Options have been rearranged
 - Implemented option to copy profile style settings from other profiles
+- Implemented style editor
 - Implemented icon selector
 - Implemented bar duplication
 - Implemented configuration for all bars